+++ DelvUI/Helpers/InputsHelper.cs
@@ -22,8 +22,10 @@ using Dalamud.Game.ClientState.Objects.Enums;
 using Dalamud.Game.ClientState.Objects.SubKinds;
 using Dalamud.Game.ClientState.Objects.Types;
 using Dalamud.Hooking;
+using Dalamud.Logging;
 using DelvUI.Config;
 using DelvUI.Interface.GeneralElements;
+using ImGuiNET;
 using Lumina.Excel;
 using System;
 using System.Diagnostics;
@@ -35,10 +37,10 @@ namespace DelvUI.Helpers
     public delegate void OnSetUIMouseoverActor(long arg1, long arg2);
     public delegate ulong OnRequestAction(long arg1, uint arg2, ulong arg3, long arg4, uint arg5, uint arg6, int arg7);
 
-    public unsafe class MouseOverHelper : IDisposable
+    public unsafe class InputsHelper : IDisposable
     {
         #region Singleton
-        private MouseOverHelper()
+        private InputsHelper()
         {
             _sheet = Plugin.DataManager.GetExcelSheet<Action>();
 
@@ -99,11 +101,11 @@ namespace DelvUI.Helpers
             OnConfigReset(ConfigurationManager.Instance);
         }
 
-        public static void Initialize() { Instance = new MouseOverHelper(); }
+        public static void Initialize() { Instance = new InputsHelper(); }
 
-        public static MouseOverHelper Instance { get; private set; } = null!;
+        public static InputsHelper Instance { get; private set; } = null!;
 
-        ~MouseOverHelper()
+        ~InputsHelper()
         {
             Dispose(false);
         }
@@ -128,6 +130,10 @@ namespace DelvUI.Helpers
 
             _requsetActionHook?.Disable();
             _requsetActionHook?.Dispose();
+
+            // give imgui the control of inputs again
+            IntPtr windowHandle = Process.GetCurrentProcess().MainWindowHandle;
+            SetWindowLongPtr(windowHandle, GWL_WNDPROC, _imguiWndProcPtr);
 
             Instance = null!;
         }
@@ -225,32 +231,48 @@ namespace DelvUI.Helpers
         }
 
         #region mouseover inputs proxy
-        public bool HandlingInputs => Target != null;
+        public bool HandlingMouseInputs => Target != null;
 
-        public bool LeftButtonClicked = false;
-        public bool RightButtonClicked = false;
+        private bool? _leftButtonClicked = null;
+        public bool LeftButtonClicked => _leftButtonClicked.HasValue ? _leftButtonClicked.Value : ImGui.GetIO().MouseClicked[0];
+
+        private bool? _rightButtonClicked = null;
+        public bool RightButtonClicked => _rightButtonClicked.HasValue ? _rightButtonClicked.Value : ImGui.GetIO().MouseClicked[1];
 
         // wnd proc detour
         // if we're "eating" inputs, we only process left and right clicks
         // any other message is passed along to the ImGui scene
         private IntPtr WndProcDetour(IntPtr hWnd, uint msg, ulong wParam, long lParam)
         {
+            // esc key press
+            if (msg == WM_KEYDOWN && wParam == 27 && ConfigurationManager.Instance.UseEscInput())
+            {
+                return (IntPtr)0;
+            }
             // eat left and right clicks?
-            if (HandlingInputs)
+            else if (HandlingMouseInputs)
             {
                 switch (msg)
                 {
+                    // mouse clicks
                     case WM_LBUTTONDOWN:
                     case WM_RBUTTONDOWN:
-                        return (IntPtr)0;
-
                     case WM_LBUTTONUP:
-                        LeftButtonClicked = true;
-                        return (IntPtr)0;
-
                     case WM_RBUTTONUP:
-                        RightButtonClicked = true;
-                        return (IntPtr)0;
+
+                        // if there's not a game window covering the cursor location
+                        // we eat the message and handle the inputs manually
+                        if (ClipRectsHelper.Instance?.IsPointClipped(ImGui.GetMousePos()) == false)
+                        {
+                            _leftButtonClicked = msg == WM_LBUTTONUP;
+                            _rightButtonClicked = msg == WM_RBUTTONUP;
+                            return (IntPtr)0;
+                        }
+
+                        // otherwise we let imgui handle the inputs
+                        _leftButtonClicked = null;
+                        _rightButtonClicked = null;
+                        break;
                 }
             }
 
@@ -260,8 +282,8 @@ namespace DelvUI.Helpers
 
         public void Update()
         {
-            LeftButtonClicked = false;
-            RightButtonClicked = false;
+            _leftButtonClicked = null;
+            _rightButtonClicked = null;
         }
 
         public delegate IntPtr WndProcDelegate(IntPtr hWnd, uint msg, ulong wParam, long lParam);
@@ -272,6 +294,7 @@ namespace DelvUI.Helpers
         [DllImport("user32.dll", EntryPoint = "CallWindowProcW")]
         public static extern long CallWindowProc(IntPtr lpPrevWndFunc, IntPtr hWnd, uint Msg, ulong wParam, long lParam);
 
+        private const uint WM_KEYDOWN = 256;
         private const uint WM_LBUTTONDOWN = 513;
         private const uint WM_LBUTTONUP = 514;
         private const uint WM_RBUTTONDOWN = 516;