+++ DelvUI/Interface/GeneralElements/HUDOptionsConfig.cs
@@ -18,8 +18,12 @@ namespace DelvUI.Interface.GeneralElements
         [Order(6, collapseWith = nameof(UseGlobalHudShift))]
         public Vector2 HudOffset = new(0, 0);
 
-        [Checkbox("Mouseover", separator = true)]
+        [Checkbox("Dim DelvUI's settings window when not focused")]
         [Order(10)]
+        public bool DimConfigWindow = false;
+
+        [Checkbox("Mouseover", separator = true)]
+        [Order(15)]
         public bool MouseoverEnabled = true;
 
         [Checkbox("Automatic Mode", help =
@@ -27,16 +31,20 @@ namespace DelvUI.Interface.GeneralElements
             "Mouseover macros or other mouseover plugins are not necessary and WON'T WORK in this mode!\n\n" +
             "When disabled: DelvUI unit frames will behave like the game's ones.\n" +
             "You'll need to use mouseover macros or other mouseover related plugins in this mode.")]
-        [Order(11, collapseWith = nameof(MouseoverEnabled))]
+        [Order(16, collapseWith = nameof(MouseoverEnabled))]
         public bool MouseoverAutomaticMode = true;
 
-        [Checkbox("Hide DelvUI outside of combat", separator = true, help = "Show in Duty-option available once enabed.")]
+        [Checkbox("Hide DelvUI outside of combat", separator = true, help = "Show in Duty and Show on Weapon Drawn-options available once enabed.")]
         [Order(20)]
         public bool HideOutsideOfCombat = false;
 
         [Checkbox("Show in duty")]
         [Order(21, collapseWith = nameof(HideOutsideOfCombat))]
         public bool ShowInDuty = false;
+
+        [Checkbox("Show on Weapon Drawn")]
+        [Order(22, collapseWith = nameof(HideOutsideOfCombat))]
+        public bool ShowOnWeaponDrawn = false;
 
         [Checkbox("Hide DelvUI in Gold Saucer")]
         [Order(25)]
@@ -46,24 +54,28 @@ namespace DelvUI.Interface.GeneralElements
         [Order(30)]
         public bool HideOnlyJobPackHudOutsideOfCombat = false;
 
-        [Checkbox("Hide Default Job Gauges", isMonitored = true, spacing = true)]
+        [Checkbox("Automatically disable HUD elements preview", help = "Enabling this will make it so all HUD elements preview modes are disabled when DelvUI's setting window is closed")]
         [Order(35)]
+        public bool AutomaticPreviewDisabling = true;
+
+        [Checkbox("Hide Default Job Gauges", isMonitored = true, spacing = true)]
+        [Order(40)]
         public bool HideDefaultJobGauges = false;
 
         [Checkbox("Hide Default Castbar", isMonitored = true)]
-        [Order(40)]
+        [Order(45)]
         public bool HideDefaultCastbar = false;
 
         [Checkbox("Hide Default Pulltimer", isMonitored = true)]
-        [Order(45)]
+        [Order(50)]
         public bool HideDefaultPulltimer = false;
 
         [Checkbox("Enable Combat Hotbars", isMonitored = true, separator = true)]
-        [Order(50)]
+        [Order(200)]
         public bool EnableCombatActionBars = false;
 
         [DynamicList("Hotbars Shown Only In Combat", "Hotbar 1", "Hotbar 2", "Hotbar 3", "Hotbar 4", "Hotbar 5", "Hotbar 6", "Hotbar 7", "Hotbar 8", "Hotbar 9", "Hotbar 10", isMonitored = true)]
-        [Order(81, collapseWith = nameof(EnableCombatActionBars))]
+        [Order(201, collapseWith = nameof(EnableCombatActionBars))]
         public List<string> CombatActionBars = new List<string>();
 
         public Vector2 CastBarOriginalPosition;