+++ DelvUI/Interface/GeneralElements/UnitFrameHud.cs
@@ -13,7 +13,7 @@ using System.Runtime.InteropServices;
 
 namespace DelvUI.Interface.GeneralElements
 {
-    public unsafe class UnitFrameHud : DraggableHudElement, IHudElementWithActor
+    public unsafe class UnitFrameHud : DraggableHudElement, IHudElementWithActor, IHudElementWithMouseOver
     {
         public UnitFrameConfig Config => (UnitFrameConfig)_config;
 
@@ -48,10 +48,20 @@ namespace DelvUI.Interface.GeneralElements
             return (new List<Vector2>() { Config.Position }, new List<Vector2>() { Config.Size });
         }
 
+        public void StopMouseover()
+        {
+            if (_wasHovering)
+            {
+                InputsHelper.Instance.Target = null;
+                _wasHovering = false;
+            }
+        }
+
         public override void DrawChildren(Vector2 origin)
         {
             if (!Config.Enabled || Actor == null)
             {
+                StopMouseover();
                 return;
             }
 
@@ -68,14 +78,14 @@ namespace DelvUI.Interface.GeneralElements
             var startPos = Utils.GetAnchoredPosition(origin + Config.Position, Config.Size, Config.Anchor);
             if (ImGui.IsMouseHoveringRect(startPos, startPos + Config.Size) && !DraggingEnabled)
             {
-                MouseOverHelper.Instance.Target = Actor;
+                InputsHelper.Instance.Target = Actor;
                 _wasHovering = true;
 
-                if (MouseOverHelper.Instance.LeftButtonClicked)
+                if (InputsHelper.Instance.LeftButtonClicked)
                 {
                     Plugin.TargetManager.SetTarget(Actor);
                 }
-                else if (MouseOverHelper.Instance.RightButtonClicked)
+                else if (InputsHelper.Instance.RightButtonClicked)
                 {
                     var agentHud = new IntPtr(Framework.Instance()->GetUiModule()->GetAgentModule()->GetAgentByInternalID(4));
                     _openContextMenuFromTarget(agentHud, Actor.Address);
@@ -83,7 +93,7 @@ namespace DelvUI.Interface.GeneralElements
             }
             else if (_wasHovering)
             {
-                MouseOverHelper.Instance.Target = null;
+                InputsHelper.Instance.Target = null;
                 _wasHovering = false;
             }
         }