+++ DelvUI/Interface/Party/PartyFramesBar.cs
@@ -122,10 +122,27 @@ namespace DelvUI.Interface.Party
             return _config.ColorsConfig.GenericRoleColor;
         }
 
+        public void StopPreview()
+        {
+            _castbarHud.StopPreview();
+            _buffsListHud.StopPreview();
+            _debuffsListHud.StopPreview();
+        }
+
+        public void StopMouseover()
+        {
+            if (_wasHovering)
+            {
+                InputsHelper.Instance.Target = null;
+                _wasHovering = false;
+            }
+        }
+
         public void Draw(Vector2 origin, ImDrawListPtr drawList, PluginConfigColor? borderColor = null)
         {
             if (!Visible || Member is null)
             {
+                StopMouseover();
                 return;
             }
 
@@ -135,11 +152,11 @@ namespace DelvUI.Interface.Party
 
             if (isHovering)
             {
-                MouseOverHelper.Instance.Target = character;
+                InputsHelper.Instance.Target = character;
                 _wasHovering = true;
 
                 // left click
-                if (MouseOverHelper.Instance.LeftButtonClicked)
+                if (InputsHelper.Instance.LeftButtonClicked)
                 {
                     // move player bar to this spot on ctrl+alt+shift click
                     if (ImGui.GetIO().KeyCtrl && ImGui.GetIO().KeyAlt && ImGui.GetIO().KeyShift)
@@ -153,14 +170,14 @@ namespace DelvUI.Interface.Party
                     }
                 }
                 // right click (context menu)
-                else if (MouseOverHelper.Instance.RightButtonClicked)
+                else if (InputsHelper.Instance.RightButtonClicked)
                 {
                     OpenContextMenuEvent?.Invoke(this);
                 }
             }
             else if (_wasHovering)
             {
-                MouseOverHelper.Instance.Target = null;
+                InputsHelper.Instance.Target = null;
                 _wasHovering = false;
             }
 
@@ -293,14 +310,10 @@ namespace DelvUI.Interface.Party
         // need to separate elements that have their own window so clipping doesn't get messy
         public void DrawElements(Vector2 origin)
         {
-            if (!Visible || Member is null)
-            {
-                return;
-            }
-
             var player = Plugin.ClientState.LocalPlayer;
-            if (player == null)
+            if (!Visible || Member is null || player == null)
             {
+                StopMouseover();
                 return;
             }
 
@@ -311,6 +324,7 @@ namespace DelvUI.Interface.Party
             {
                 var parentPos = Utils.GetAnchoredPosition(Position, -_config.Size, _manaBarConfig.HealthBarAnchor);
                 _manaBarHud.Actor = character;
+                _manaBarHud.PartyMember = Member;
                 _manaBarHud.Draw(parentPos);
             }
 
@@ -348,19 +362,7 @@ namespace DelvUI.Interface.Party
             }
 
             // health label
-            if (character == null)
-            {
-                string oldText = _config.HealthLabelConfig.GetText();
-                _config.HealthLabelConfig.SetText(Member.HP.ToString());
-
-                _healthLabelHud.Draw(Position, _config.Size, character);
-
-                _config.HealthLabelConfig.SetText(oldText);
-            }
-            else
-            {
-                _healthLabelHud.Draw(Position, _config.Size, character);
-            }
+            _healthLabelHud.Draw(Position, _config.Size, character, null, Member.HP, Member.MaxHP);
 
             // order
             if (character == null || character?.ObjectKind != ObjectKind.BattleNpc)